+++ R/build_scaledic_skeleton.R
@@ -5,10 +5,12 @@
 #' @return Writes an Excel file with an empty template of a dic file.
 #' @export
 
-build_scaledic_skeleton <- function(filename = "dic_template.xlsx") {
-  out <- matrix(NA, ncol = length(.dic_file) - 1)
+build_scaledic_skeleton <- function(filename = "dic_template.xlsx", nrows = 0) {
+  out <- matrix(NA, ncol = length(.dic_file) - 1, nrow = nrows)
   out <- as.data.frame(out)
   names(out) <- unlist(.dic_file[-1])
+
+  if (filename == "") return(out)
   openxlsx::write.xlsx(out, filename)
   cat(filename, " written at ", getwd(), "\n")
 }