+++ R/print.dic.R
@@ -1,0 +1,11 @@
+#' Print dic infos
+#'
+#' @param x A variable with dic infos
+#' @param ... Further parameters for the print function
+#'
+#' @return Dic infos of x
+#' @export
+print.dic <- function(x, ...) {
+  dic(x)
+
+}